refactor: move custom Suppliers implementation to a utility package
